refactor password migrate scripts
