+++ package.json
@@ -34,7 +34,7 @@
     "seed": "pnpm clear:db && pnpm exec ts-node packages/scripts/src/seed.ts",
     "clear:uploads": "pnpm exec ts-node packages/scripts/src/clear-uploads-folder.ts",
     "clear:db": "pnpm exec ts-node packages/scripts/src/clear-db.ts",
-    "migrate:password": "node ./packages/scripts/src/password-migrate.mjs && node ./packages/scripts/src/password-inject.mjs"
+    "migrate:password": "node ./packages/scripts/src/password-migrate.mjs"
   },
   "author": "ReebA Team",
   "license": "MIT",
